requirements.txt
streamlit==1.36.0
pandas==2.1.4
openpyxl==3.11.0
xlrd==2.0.1
langchain==0.1.10
langchain-community==0.0.25
requests==2.31.0
python-dotenv==1.0.0
markdown==3.5.1
Pillow==10.1.0
pyarrow==14.0.1
